requirements.txt
decorator==4.0.2
Django==1.8.4
django-filter==0.11.0
djangorestframework==3.2.3
ipython==4.0.0
ipython-genutils==0.1.0
Markdown==2.6.2
mongoengine==0.10.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
pymongo==3.0.3
simplegeneric==0.8.1
traitlets==4.0.0
